wg_dashboard_backend/requirements.txt
pydantic
fastapi
aiofiles
aiosqlite
sqlalchemy
databases
PyJWT
passlib
bcrypt
python-multipart
jinja2
sqlalchemy_utils
sqlalchemy-migrate
requests
